Moved and renamed service logger.
