+++ rxandroidble/src/test/groovy/com/polidea/rxandroidble2/internal/connection/RxBleConnectionTest.groovy
@@ -10,7 +10,7 @@ import com.polidea.rxandroidble2.internal.operations.OperationsProviderImpl
 import com.polidea.rxandroidble2.internal.operations.ReadRssiOperation
 import com.polidea.rxandroidble2.internal.util.ByteAssociation
 import com.polidea.rxandroidble2.internal.util.MockOperationTimeoutConfiguration
-import com.polidea.rxandroidble2.internal.util.RxBleServicesLogger
+import com.polidea.rxandroidble2.internal.logger.LoggerUtilBluetoothServices
 import io.reactivex.Completable
 import io.reactivex.Observable
 import io.reactivex.Scheduler
@@ -42,7 +42,7 @@ class RxBleConnectionTest extends Specification {
     def illegalOperationChecker = Mock IllegalOperationChecker
     def testScheduler = new TestScheduler()
     def timeoutConfig = new MockOperationTimeoutConfiguration(testScheduler)
-    def operationsProviderMock = new OperationsProviderImpl(gattCallback, bluetoothGattMock, Mock(RxBleServicesLogger),
+    def operationsProviderMock = new OperationsProviderImpl(gattCallback, bluetoothGattMock, Mock(LoggerUtilBluetoothServices),
             timeoutConfig, testScheduler, testScheduler,
             { new ReadRssiOperation(gattCallback, bluetoothGattMock, timeoutConfig) })
     def notificationAndIndicationManagerMock = Mock NotificationAndIndicationManager
